refactor(): lowercase file names for Followers and Following components
